backend/requirements.txt
# Frameworks y extensiones principales
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-SSE==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==4.6.0

# Utilidades
requests==2.31.0
Werkzeug==2.3.7
marshmallow==3.20.1
validators==0.22.0
pydantic==2.4.2
gunicorn==21.2.0

# Herramientas de desarrollo
pytest==7.4.2
black==23.9.1
flake8==6.1.0